+++ {{cookiecutter.project_name}}/tests/test_cli.py
@@ -1,29 +1,24 @@
 """Tests for `{{ cookiecutter.project_slug }}`.cli module."""
+from typing import List
+
+import pytest
 from click.testing import CliRunner
 
 import {{ cookiecutter.project_slug }}
 from {{ cookiecutter.project_slug }} import cli
 
 
-def test_command_line_interface() -> None:
+@pytest.mark.parametrize(
+    "options,expected",
+    [
+        ([], "{{ cookiecutter.project_slug }}.cli.main"),
+        (["--help"], "Usage: main [OPTIONS]"),
+        (["--version"], f"main, version { {{ cookiecutter.project_slug }}.__version__ }\n"),
+    ],
+)
+def test_command_line_interface(options: List[str], expected: str) -> None:
     """Test the CLI."""
     runner = CliRunner()
-    result = runner.invoke(cli.main)
+    result = runner.invoke(cli.main, options)
     assert result.exit_code == 0
-    assert "{{ cookiecutter.project_slug }}.cli.main" in result.output
-
-
-def test_command_line_interface_help() -> None:
-    """Test the CLI help option."""
-    runner = CliRunner()
-    help_result = runner.invoke(cli.main, ["--help"])
-    assert help_result.exit_code == 0
-    assert "Usage: main [OPTIONS]" in help_result.output
-
-
-def test_command_line_interface_version() -> None:
-    """Test the CLI version option."""
-    runner = CliRunner()
-    help_result = runner.invoke(cli.main, ["--version"])
-    assert help_result.exit_code == 0
-    assert f"main, version { {{ cookiecutter.project_slug }}.__version__ }\n" == help_result.output
+    assert expected in result.output